requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
jinja2>=3.1.3
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Media Processing
ffmpeg-python>=0.2.0
Pillow>=10.4.0
pillow-heif>=1.0.0
pypdf>=4.0.1

# AI Integration
google-genai>=0.4.0
httpx>=0.27.0

# Utilities
aiofiles>=23.2.1
